Reverts cackleberry name change
About The Pull Request

As the title says, changes the name of the cackleberries from egg back to cackleberry

Why It's Good For The Game

I think it's very "soulful" that eggs are called cackleberries in the world of Vanderlin.
Also consistency with the ingame universal accent, if you try to say "egg" ingame it will convert to "cackleberry"

Pre-Merge Checklist

You tested this on a local server.
This code did not runtime during testing.
You documented all of your changes.