(SOUND) Adds a Baothan combat theme (#2203)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

Adds the song 'Love Within You (Rough Mix)' by T87-Sulfurhead to the
following roles, so long as they are Baothan:

Paladin
Cleric
Minor Noble
Consort

## Why It's Good For The Game

Baotha's been needing a theme for a long while, god damn

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->