Adds more migrant roles II (#2181)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

Adds three more migrant classes:

- BODYGUARD
An unarmed/knife user with the goals of finding an employer, and not
working with the mercenary's guild.


![image](https://github.com/user-attachments/assets/d4b9c94d-2de1-41db-9529-be1c9b5813fe)

- CHIRURGEON
A back-alley doctor shadier than other physickers. Has some survival
skills to encourage setting up a practice in the middle of bumfuck
nowhere.


![image](https://github.com/user-attachments/assets/b286b288-87dd-4eb3-8036-f7711f4bf550)


- ESCAPED PRISONER
Hunter without any of the gear. Low constitution, survival skills. A
former noble, outlawed by default. Don't get caught.


![image](https://github.com/user-attachments/assets/3f275fa7-3ae4-44c8-b9cd-ea9036703c1c)


Also bumped up the chances of a Daywalker appearing up from 1 -> 3


## Why It's Good For The Game

We're severely lacking migrant variety, and many of the migrant roles
are big-ticket items that require several players. Most of the roles in
this PR can just be mimicked by selecting a similar adventuring class,
but sometimes players need a push to do a gimmick.

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->