optimizations  (#2060)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request
needs ma2html for new family tree rendering so
ports: https://github.com/DaedalusDock/daedalusdock/pull/1029

ook wanted to test shadowcasting again which modifys the code from:

https://github.com/DaedalusDock/daedalusdock/compare/master...Kapu1178:daedalusdock:shadowcasting#diff-b4e5189217aa0181c07326f255c6107c40012f84dbf818a739b7cddfcd84f0ee
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

## Why It's Good For The Game

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [ ] You tested this on a local server.
- [ ] This code did not runtime during testing.
- [ ] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->