Necra's Miracles Changes #2123 (#2211)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request


[github.com/Monkestation/Vanderlin/pull/2123/files](https://github.com/Monkestation/Vanderlin/pull/2123#issue-3097513184)

Tyger destroyed their repo, they asked me to set up the PR's in their
name. Follow to the closed PR and ask them on discord if you have code
comments. I do not know anything about this.

## Why It's Good For The Game

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [ ] You tested this on a local server.
- [ ] This code did not runtime during testing.
- [ ] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->