Add Duelists from Valoria. (#2224)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request
Title, This PR add a new type of mercs that are just duelists from
Valoria.
They get a unique cape and hat and some clothes.

![image](https://github.com/user-attachments/assets/8882ed07-ddb8-4fe4-b96c-3077011c3837)

![image](https://github.com/user-attachments/assets/ec48475a-e17b-42b8-81f9-e4ed2263f0ee)

![image](https://github.com/user-attachments/assets/ae2db489-2057-4248-ad27-fb33083061bb)


![image](https://github.com/user-attachments/assets/7eb71f6b-c8f4-416e-acd5-92ee33ba0e92)

their medal when they receive a token.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

## Why It's Good For The Game
More mercs good, Huge thanks to @7erracotta for the sprites
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->