Adds Save Log button to chat (#2134)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->
Adds a "save chat log" button
## Why It's Good For The Game

Are you tried of going "Wait shit I should save that moment" and then
having to spend a minute or so away from the game to save it because you
need to start your recording software or your screen shotting and saving
somewhere?

Do you literally just want to dump the chat so you can save it to
remember "oh yea that's what I did".

Well now you can. I enabled it. I don't know why it was disabled in the
first place.

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->