Fixes theives cant, and tongueless speech
Thieves cant can now be spoken with a tongue. It also now works as intended. People without tongues can now speak, but with the garbling handled in-game, no more making a strange sound, and more Haaaalps. Adds more sign language verbs to the list of possible ones if you don't understand thieves cant. I had to add the tongueless speech flag to all the languages (besides celestial because I like to imagine the language of the Gods can't be understood at all if spoken without a tongue).